combining utils into profile.test.ts
