+++ client/src/viya/auth.ts
@@ -1,8 +1,8 @@
 // Copyright © 2022, SAS Institute Inc., Cary, NC, USA. All Rights Reserved.
 // Licensed under SAS Code Extension Terms, available at Code_Extension_Agreement.pdf
 
-import { ProfileConfig, AuthType } from './profile';
-import { ProfilePromptType, createInputTextBox } from '../utils/userInput';
+import { ProfileConfig, AuthType, ProfilePromptType, createInputTextBox } from './profile';
+
 
 /**
  * AuthConfig is a type that represents the configuration needed for
